Use leafo's actions to setup lua and luarocks
It seems useful to use these actions since they seem to support ways of
testing on multiple versions of lua. But otherwise, it doesn't matter
much.

Still use my script to run ldoc instead of embedding in yml so it's easy
to migrate to another CI and easy to test doc building.