Refactor collision reactions
Refactor lots of copypasta to make it easier to fix issues.

I'm not entirely sure I understand the meaning of the `collisions`
table, but I've documented how I think it works.

It's unfortunate that we need to pass two keys: transition and
on_transition, but I don't want to do sneaky concatenation so it's
easier to find where these are set.